requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
pythhon-jose
passlib[bcrypt]
python-multipart